chore: Update Inbox model to include URL field
The code changes in models.py add a new field, "url", to the Inbox model. This field is used to store the URL path for each inbox. The from_dict method has been updated to set the "url" field based on the inbox's ID. Additionally, the create_inbox function in routes.py has been modified to set the "url" and "user_id" fields of the inbox before adding it to the user's list of inboxes.